fix: changed min width and added queries to all call-components that need it
